pyproject.toml
[project]
name = "skillgrade"
version = "0.2.0"
description = "Skill lifecycle management: create, evaluate, and evolve AI agent skills"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Minko Gechev"}
]
keywords = ["ai", "agent", "skills", "evaluation", "benchmark", "llm", "evolution"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core dependencies
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "langgraph-sdk>=0.1.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "orjson>=3.10.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    # skillevol dependencies
    "typer>=0.12.0",
    "rich>=13.7.0",
    "anthropic>=0.25.0",
    "openai>=1.12.0",
    # skillforge dependencies
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
skillgrade = "skillgrade.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
